+++ nextgisweb/vector_layer/test/ref_ddl/add_fields.v=true.sql
@@ -1,0 +1,5 @@
+ALTER TABLE ct ADD COLUMN fld_i INT,
+    ADD COLUMN fld_d DATE;
+
+ALTER TABLE ht ADD COLUMN fld_i INT,
+    ADD COLUMN fld_d DATE;